WIP: a11y refactoring to remove decorator
