+++ addons/a11y/src/constants.ts
@@ -2,11 +2,3 @@ export const ADDON_ID = 'storybook/a11y';
 export const PANEL_ID = `${ADDON_ID}/panel`;
 export const PARAM_KEY = `a11y`;
 export const IFRAME = 'iframe';
-export const ADD_ELEMENT = 'ADD_ELEMENT';
-export const CLEAR_ELEMENTS = 'CLEAR_ELEMENTS';
-const RESULT = `${ADDON_ID}/result`;
-const REQUEST = `${ADDON_ID}/request`;
-const ERROR = `${ADDON_ID}/error`;
-const MANUAL = `${ADDON_ID}/manual`;
-
-export const EVENTS = { RESULT, REQUEST, ERROR, MANUAL };
